# packages/kanban-viewer/openapi.yaml
openapi: 3.1.0
info:
  title: A(i)-Team Kanban Viewer API
  version: 1.0.0
  description: |
    REST API for the A(i)-Team Kanban Viewer. Manages work items, missions,
    agent lifecycle, and real-time board events for the TDD pipeline.

    ## Project Isolation
    All requests (except `/api/projects`) require the `X-Project-ID` header.
    Projects are created automatically on first use via `POST /api/projects`.

    ## Error Format
    All errors return a consistent envelope:
    ```json
    { "success": false, "error": { "code": "ERROR_CODE", "message": "...", "details": {} } }
    ```
servers:
  - url: http://localhost:3000
    description: Local development

# ─── Reusable components ─────────────────────────────────────────────────────
components:
  parameters:
    ProjectId:
      name: X-Project-ID
      in: header
      required: true
      description: Project identifier for multi-tenant isolation
      schema:
        type: string
        example: my-project
  schemas:
...
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum:
                - VALIDATION_ERROR
                - INVALID_TRANSITION
                - WIP_LIMIT_EXCEEDED
                - INVALID_STAGE
                - DEPENDENCY_CYCLE
                - DEPENDENCIES_NOT_MET
                - OUTPUT_COLLISION
                - CLAIM_CONFLICT
                - ITEM_CLAIMED
                - NOT_CLAIMED
                - CLAIM_MISMATCH
                - ITEM_NOT_FOUND
                - MISSION_NOT_FOUND
                - NO_ACTIVE_MISSION
                - INVALID_MISSION_STATE
                - PROJECT_NOT_FOUND
                - STAGE_NOT_FOUND
                - CONFLICT
                - UNAUTHORIZED
                - SERVER_ERROR
                - DATABASE_ERROR
            message:
              type: string
            details:
              description: Optional structured context for the error
...
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

# ─── Paths ────────────────────────────────────────────────────────────────────
paths:
...
  /api/tuning/proposals/{id}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - name: id
        in: path
        required: true
        schema:
          type: integer

    patch:
      operationId: updateTuningProposal
      summary: Apply a tuning verb (accept|edit|merge) to an existing proposal
      description: |
        `accept`/`edit` promote the proposal to status='accepted', gated by
        the same corroboration check as create (422 NOT_CORROBORATED unless
        every linked fingerprint is corroborated); `edit` may also re-target
        `proposalText`/`targetSurface`/`altitude`. `merge` de-aliases another
        linked fingerprint into `mergeIntoFingerprint` and is not gated by
        corroboration, but is rejected with 422 SAME_MISSION_MERGE if a
        source fingerprint shares a mission with the target. `reject` and
        `demote` no longer exist as verbs here — they were collapsed into the
        durable per-fingerprint `POST /api/tuning/fingerprints/{slug}/defer`.
      tags: [tuning]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [verb]
              properties:
                verb:
                  type: string
                  enum: [accept, edit, merge]
                proposalText:
                  type: string
                  description: Amended proposal text (edit)
                targetSurface:
                  type: string
                  description: Re-targeted surface (edit)
                altitude:
                  type: string
                  enum: [skill-text, agent-prompt, hook]
                  description: Re-targeted altitude (edit)
                mergeIntoFingerprint:
                  type: string
                  description: Target fingerprint slug to merge into (merge, required for that verb)
      responses:
        '200':
          description: Proposal updated
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: PROPOSAL_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: NOT_CORROBORATED (accept/edit) · SAME_MISSION_MERGE (merge)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /api/tuning/fingerprints/{slug}/defer:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - name: slug
        in: path
        required: true
        schema:
          type: string

    post:
      operationId: deferTuningFingerprint
      summary: Durable "not now" watermark for a fingerprint
      description: |
        No request body. Sets `Fingerprint.deferredAtMissions` to the
        fingerprint's current `distinctMissions`. The fingerprint reports
        `actionable: false` on `GET /api/tuning/candidates` until
        `distinctMissions` climbs DEFER_MARGIN past that watermark — this
        replaces the removed `reject`/`demote` verbs.
      tags: [tuning]
      responses:
        '200':
          description: Fingerprint deferred
          content:
            application/json:
              schema:
                type: object
                required: [success, data]
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    type: object
                    required: [slug, deferredAtMissions, distinctMissions]
                    properties:
                      slug:
                        type: string
                      deferredAtMissions:
                        type: integer
                      distinctMissions:
                        type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: FINGERPRINT_NOT_FOUND
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/ServerError'

  # ── Missions ─────────────────────────────────────────────────────────────────
...